Separate the extension and add tests
